refactor :: 현재 유저 조회 패키지 경로 변경
