+++ motung-i-api/src/main/kotlin/kr/motung_i/backend/domain/music/usecase/CreateMusicUsecase.kt
@@ -1,7 +1,7 @@
 package kr.motung_i.backend.domain.music.usecase
 
 import kr.motung_i.backend.domain.music.presentation.dto.request.CreateMusicRequest
-import kr.motung_i.backend.domain.user.presentation.usecase.FetchCurrentUserUsecase
+import kr.motung_i.backend.domain.user.usecase.FetchCurrentUserUsecase
 import kr.motung_i.backend.persistence.music.entity.Music
 import kr.motung_i.backend.persistence.music.repository.MusicRepository
 import org.springframework.stereotype.Service
