+++ motung-i-api/src/main/kotlin/kr/motung_i/backend/domain/user/usecase/FetchCurrentUserUsecase.kt
@@ -1,4 +1,4 @@
-package kr.motung_i.backend.domain.user.presentation.usecase
+package kr.motung_i.backend.domain.user.usecase
 
 import kr.motung_i.backend.persistence.user.entity.User
 import kr.motung_i.backend.persistence.user.repository.UserRepository
